pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyperliquid-requester"
version = "0.1.0"
description = "Agent-market assisted Avellaneda-Stoikov requester for Hyperliquid"
readme = "README.md"
authors = [{ name = "Grouplang" }]
license = { file = "LICENSE" }
requires-python = ">=3.9"
dependencies = [
    "requests>=2.32.0",
    "hyperliquid-python-sdk>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "python-dotenv>=1.0.1",
]

[project.scripts]
hyperliquid-requester = "hyperliquid_requester.market_maker:main"
